// src/app/interfaces/dons.interface.ts
export enum DeliveryMethod {
    DROP_OFF = 'DROP_OFF',
    PICK_UP = 'PICK_UP'
}

export interface Dons {
    idDons?: number;                      // Optional, since it's generated by the backend
    nomDoneur: string;                     // Donor's first name
    prenomDoneur: string;                  // Donor's last name
    donorEmail:String
    quantite: number;                      // Quantity of donation
    donationDate?: Date;                   // Optional, will be set by the backend
    scheduledDateTime?: Date;              // Optional, scheduled date and time for delivery or pick-up
    associationValidated: boolean;         // Whether the donation is validated by the association
    donorNote?: string;                    // Optional note from the donor
    deliveryMethod: DeliveryMethod;        // Method of donation delivery ('DROP_OFF' | 'PICK_UP')
    donorAddress?: string;                 // Optional address if pick-up is chosen
    idDonation: number;                    // The ID of the donation being contributed to
    subscriberId?: number;                 // Optional, if the donor is a registered subscriber
}